+++ packages/svg-sprites/test/vite/main.ts
@@ -1,5 +1,5 @@
-import "@sv/svg-sprites/svg-icon";
-import { svg, blob } from "@sv/svg-sprites/sheet";
+import "@atrium-ui/svg-sprites/svg-icon";
+import { svg, blob } from "@atrium-ui/svg-sprites/sheet";
 
 window.addEventListener("DOMContentLoaded", async () => {
   console.info(svg());