+++ packages/svg-sprites/test/vite/vite.config.ts
@@ -1,5 +1,5 @@
 import path from "node:path";
-import svgSprite from "@sv/svg-sprites/vite";
+import svgSprite from "@atrium-ui/svg-sprites/vite";
 import { defineConfig } from "vite";
 
 export default defineConfig({
@@ -9,7 +9,7 @@ export default defineConfig({
       dir: [
         path.resolve("../assets/icons/**/*.svg"),
         path.resolve("../assets/icons2/*.svg"),
-        `${path.dirname(require.resolve("@sv/icons"))}/*.svg`,
+        `${path.dirname(require.resolve("@atrium-ui/icons"))}/*.svg`,
       ],
       transform: (code, file) => {
         console.warn("transform", file);