+++ packages/components/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @sv/components
+# @atrium-ui/components
 
 ## 1.8.1 (2025-03-01)