+++ packages/components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sv/components
 
+## 0.1.0-dev.1 (2025-10-17)
+
+### Breaking Changes
+
+- Move chart and animation elemnts into own package
+
+## 0.1.0-dev.0 (2025-10-16)
+
+### Breaking Changes
+
+- Move chart and animation elemnts into own package
+
 ## 1.8.1 (2025-03-01)
 
 ### Fixes